backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.1
django-cors-headers==4.3.0
django-filter==23.3
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
environ==1.0
Markdown==3.5.1
psycopg2==2.9.9
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
